reactor-example: Enable add button only if the inputs are valid
The add button is enabled only if the input fields corresponding to required
attributes have all been fulfilled, and all fulfilled input fields contain valid
content (can be converted to the required class).

However, as of now, it's still possible to create a generic with a null value,
and the add button is enabled even if some required links are missing.