gs-reactor: Tag: Delete initProperty(String propertyName, T initialValue)
If a property is created in a tag corresponding to several models, this method
puts a reference to the same object in the properties stored in each model,
whereas initProperty(String propertyName, Function<M, T> getInitialValue) uses
different references.
This matters if the property contains, for instance, a List (or a Map): if the
property is initialized with
    initProperty(name, new ArrayList<...>());
all the models will contain the same list, which is not the desired behavior.
With
    initProperty(name, model -> new ArrayList<...>());
the models contain different lists.

Force the use of a Function argument to avoid this problem.