GSEditor: Display GSCellAdder iff it’s possible to add a link
Display GSCellAdder if, and only if, there is no link/holder for the
relation/attribute, or if the attribute/relation doesn’t have the
property/singular constraint enabled.

In the click action, reset the holders’ values or the selections *before*
creating the holder, because it causes the firing of value changed events, and
if GSCellAdder disappears after the holder’s creation, this causes null pointer
exceptions.