gs-reactor: Add pos parameter to annotations with path
In annotation where a path { class1, ..., classN } can be given, a pos argument
consisting of an array of integers { pos1, ..., posN } (starting at 0) can be
given. Then the annotation is applied only to the descendant tag that’s the
posN-th child of a class extending classN of its parent, which is itself the
posN-1-th child of a class extending classN-1 of its parent, etc.

Example:
@ReactorDependencies({ HtmlLi.class, HtmlLi.class })
@SetText(path = HtmlLi.class, pos = 0, value = "First item")
@SetText(path = HtmlLi.class, pos = 1, value = "Second item")