gs-reactor: Fix issues in editor
Initialize checkboxes in init() method and not in the constructor, because the
type “checkbox” is set by an annotation that is treated after the constructor’s
call. If the checked property is set on an input of type text, it is ignored.

Correctly display titles in the editor: gs-reactor sometimes tried to display
a checkbox.

Use the same background-color as for attributes in the edition of an instance’s
name and components.