gs-reactor: Make @Switch’s behavior more intuitive with @ForEach
Previously, if there was both a metabinding (e.g. @ForEach) and an @Switch on
the same tag, the methods defined in @Switch were called on the parent context
rather than on each of the subcontexts defined by the metabinding, which
prevented using @Switch to filter the subContexts. @Switch’s behavior is now the
one expected.