gs-reactor: Replace context properties with attributes with an Object value
It’s now possible to store anything as a name-value pair associated to a tag in
a context, and not only ObservableValues.
Refactor the functions related to properties : those that take a context as
a parameter don’t need a Function<Context, T> as a value provider, they can take
the value directly.