gs-watch: fixed several bug that occured during the OCR
The OCR and the input of the results in the Engine can now be done safely. Previously, it was working when launched from the same JVM, but failed when launched from different verticles. The serialization of the lambdas expressions was particularly problematic. To circumvent it, all the lambdas were moved to an enum (ImgFilterInstance).