gs-ir: added ActionPersistentVerticle to factorize code for ActionVerticle requiring an engine
The code that was executed in the executeBlocking block in ActionVerticle was extracted to a separate
protected function. This function is overriden in the ActionPersistentVerticle to be executed in a
cache.safeConsum block.
Moreover, the engine and the cache were factorized in the abstract class, as well as the constructor.