gs-ir; gs-cv: Attempts to fix issues with worker verticle in gs-ir
The WorkerVerticle class was updated to force the user to implement start(), and to undeploy the
worker verticle as soon as the task is complete.
In ComputeBestTextPerZone, the code was updated to prevent flushing the cache with
engine.getCurrentCache.flush() (the flush is performed from gs-reactor).
The code in ContextActionCustom was refactored to prevent undesired behaviors. As an example,
when a worker verticle is called after a click on a button, the button is de-activated while the
task is performed.