gs-cv: changed the way a Zone is identified
A Zone object will now be identified with a String (UID, unique identifier) stored as the value of the
zoneinstance. A ZoneNum class was added in order to identify a zone by its number (only for display).
The value of the instance is a unique ID, computed from its coordinates (actually, from the String
representation of its Rect object). Generic System's model will be updated accordingly to accomodate
this change.