gs-ir: Fix a list extractor
The list extractor had an unused comparator that always returned 0. With the
previous implementation this caused no problem, but
Snapshot.getIndexedElements() uses an ordered Set internally. If compareTo
returns 0 for two elements, they are considered equal. Use the correct sorting
on the generic’s value instead. This makes distincByKey redundant since generics
with the same value will be considered equal because of the comparator.