gs-cv: added new automated methods in OCRPlastyScorer
A new class Results was created in order to store the results of a computation.
This class contains the duration of the computation, the final score (similarity) and
the original and corrected Strings.
A new method was created to automatically compute the correction for a given String,
using all the algorithms defined in OCRPlasty.RANSAC enum.