gs-cv: added a method to remove overlaps in fields (fields included in other fields)
A first attempt is proposed and consists in merging fields when their common area is > 0.5
and removing the biggest field otherwise (i.e., the field with the bigger rect).
During the merge, the 'receiving' field's rectangle is updated with a rectangle whose coordinates
are the mean of the two rectangles.