gs-cv: Fields: changed the implementation of merge
Instead of copying the fields and replacing them with a list of fields created from the new
rectangles, we loop over the list of rects to find the matching fields, then merge the informations
(create a copy of the field, update the rectangle) and we delete the old field from the list.
In the case no match was found, a new Field is created.