gs-reinforcer: Improve computation of affine transformation
Instead of computing a transformation for each match, we use a method inspired
by the RANSAC algorithm, but different from it because we don’t choose the
transformation most matches agree on but the transformation that leads to the
smallest error between the transformed Labels and the Labels to align them with.

For each match, we compute the transformation. Then we compare it to the
transformations obtained from other matches. If they are close enough, we
average the parameters before computing the error.